server/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pandas
xarray
scipy
netcdf4
numpy
folium
faiss-cpu

requests
